# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PeacePay API
  description: |
    PeacePay - منصة الدفع الآمن للتجارة الإلكترونية في مصر
    
    Secure Escrow Payment Platform for E-commerce in Egypt
    
    ## Authentication
    All protected endpoints require a Bearer token obtained from the login endpoint.
    ```
    Authorization: Bearer {token}
    ```
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true|false,
      "message": "رسالة توضيحية",
      "data": { ... },
      "errors": { ... }  // Only on validation errors
    }
    ```
    
    ## Error Codes
    - `400` - Bad Request
    - `401` - Unauthorized
    - `403` - Forbidden
    - `404` - Not Found
    - `422` - Validation Error
    - `429` - Too Many Requests
    - `500` - Server Error
    
  version: 1.0.0
  contact:
    name: PeacePay Support
    email: support@peacepay.eg
  license:
    name: Proprietary
servers:
  - url: https://api.peacepay.eg/api/v1
    description: Production Server
  - url: https://staging-api.peacepay.eg/api/v1
    description: Staging Server
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP code
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code, type]
              properties:
                phone:
                  type: string
                  example: "01012345678"
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: "123456"
                type:
                  type: string
                  enum: [registration, login, password_reset]
      responses:
        '200':
          description: OTP verified successfully
        '422':
          description: Invalid or expired OTP
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...